BE-24 Moved listing site processing related functions to another file
